backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
opencv-python==4.10.0.84
pyzbar==0.1.9
rembg==2.0.62
numpy==1.26.4
gspread==6.1.4
google-auth==2.37.0
